requirements.txt
streamlit>=1.32.0
networkx>=3.2
pyvis>=0.3.2
pymupdf>=1.23.0
python-docx>=1.1.0
google-generativeai>=0.5.0
pandas>=2.0.0
python-dotenv>=1.0.0
